Recombine tests into single module to reduce memory usage
